feat: add to wishlist
