+++ src/pages/CartPage/Cart.jsx
@@ -1,62 +1,16 @@
 import "./Cart.css";
-import axios from "axios";
 import {useCart} from "../../hooks/useCart";
-import {useAuth} from "../../hooks/useAuth";
 import NavBar from "../../components/NavBar/NavBar"
-import { ToastHandler } from "../../utils/toastify";
-import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import { DataCard } from "./cart-component/DataCard";
+
 
 export const Cart=()=>{
-    const {state,dispatch}= useCart();
-    const cart=[...state.cartData];
-    const {token}= useAuth();
-    const encodedToken=token;
+    const {state}= useCart();
+    const cart=state.cartData;
     const navigate= useNavigate();
-
-    const removeHandler = async (productId) => {
-      try{
-        var response = await axios.delete(
-          `/api/user/cart/${productId}`,
-          {
-              headers:{
-                authorization: encodedToken,
-            },
-          }
-        );
-        dispatch({type:"ADD_TO_CART", payload:response.data.cart});
-        ToastHandler("warn","Item removed successfully");
-      }
-      catch(error)
-      {
-        console.log(error);
-      }
-     
-    
-    };
   
-    const increaseCountHandler = (id) => {
-      dispatch({ type:"PRODUCT_COUNT", payload:
-        cart.map((item) =>
-          item._id === id ? { ...item, quantity: item.quantity + 1 } : item
-        )}
-      );
-    };
-  
-    const decreaseCountHandler = (id) => {
-        dispatch({ type:"PRODUCT_COUNT", payload:
-        cart.map((item) =>
-          item._id === id ? { ...item, quantity:item.quantity>1? item.quantity - 1:1 } : item
-        )}
-      );
-    };
-
-    useEffect(()=>{
-        if(!token)
-         navigate("/login");
-    },[])
-  
-    return (
+return (
         <>
         <NavBar/>
       <main className="main">
@@ -67,54 +21,11 @@ export const Cart=()=>{
           <div className="cart-list flex-col">
             {cart.map((item) => {
               return (
-                <div className="cart-horizontal flex-row">
-                  <div className="cart-image">
-                    <img src={item.image} alt="sneaker" className="image" />
-                  </div>
-                  <div className="cart-content flex-col cart-display">
-                    <p className="bigger-cart-text">
-                      <strong classNameName="text-center">{item.title}</strong>
-                    </p>
-                    <p className="item-price">
-                      <strong>
-                        {item.price}{" "}
-                        <span className="deleted-price">
-                          <del>Rs. {item.originalPrice}</del>
-                        </span>
-                      </strong>
-                    </p>
-                    <span className="discount">{item.discount}% OFF</span>
-                    <section className="product-count flex-row">
-                      <p>Quantity:</p>
-                      <p
-                        style={{ cursor: "pointer" }}
-                        onClick={() => decreaseCountHandler(item._id)}
-                      >
-                        <i className="fas fa-minus-circle"></i>
-                      </p>
-                      <div className="count">
-                        <p>{item.quantity}</p>
-                      </div>
-                      <p
-                        style={{ cursor: "pointer" }}
-                        onClick={() => increaseCountHandler(item._id)}
-                      >
-                        <i className="fas fa-plus-circle"></i>
-                      </p>
-                    </section>
-                    <footer>
-                    
-                      <button
-                        className="btn action-btn btn-icon  flex-row"
-                        onClick={() => removeHandler(item._id)}
-                      >
-                        Remove from Cart
-                      </button>
-                    </footer>
-                  </div>
-                </div>
-              );
-            })}
+                <DataCard
+                item={item}
+                />
+                ); 
+             })} 
           </div>
          
           {cart.length > 0 && (
